+++ custom_components/neohub_control/const.py
@@ -1,0 +1,9 @@
+"""Constants for the NeoHub Control Home Assistant integration helper."""
+
+DOMAIN = "neohub_control"
+ADDON_REPOSITORY_URL = "https://github.com/GitDakky/NeoHub-Control"
+ADDON_STORE_REDIRECT_URL = (
+    "https://my.home-assistant.io/redirect/supervisor_add_addon_repository/"
+    "?repository_url=https%3A%2F%2Fgithub.com%2FGitDakky%2FNeoHub-Control"
+)
+ISSUE_TRACKER_URL = "https://github.com/GitDakky/NeoHub-Control/issues"